% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.life.table.R
\name{.compute_survival_rates}
\alias{.compute_survival_rates}
\title{Compute Survival Rates}
\usage{
.compute_survival_rates(combined_df, strata_value = 1)
}
\arguments{
\item{combined_df}{A dataframe with 'time', 'n.risk', and 'n.event' columns.
The data should be already grouped by time.}

\item{strata_value}{A numeric value to fill the 'strata' column. Defaults to 1.}
}
\value{
A dataframe with added columns 'survival_rate', 'surv', 'upper', 'lower',
        'strata', and 'n.censor'.
}
\description{
This function computes survival rates for each time period, as well as cumulative
survival rates, based on grouped data in the input dataframe.
It then enriches the input dataframe with these calculated values and other parameters.
}
\examples{
df <- data.frame(time = c(1,2,3), n.risk = c(5,10,15), n.event = c(1,2,3))
compute_survival_rates(df, strata_value = 2)
}
