% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.coxph.R
\name{ds.coxph}
\alias{ds.coxph}
\title{Distributed Cox Proportional Hazards Model}
\usage{
ds.coxph(df, expl_vars, time_col, censor_col, datasources = NULL)
}
\arguments{
\item{df}{the name of the data frame that contains the variables to be used 
in the Cox PH model.}

\item{expl_vars}{list of the explanatory variables (covariates) that are
included as columns in \code{df}}

\item{time_col}{name of the column in \code{df} that contains the event/censor times}

\item{censor_col}{name of the column in \code{df} that explains whether an event 
occurred or the patient was censored}

\item{datasources}{a list of \code{\link{DSConnection-class}} 
objects obtained after login. If the \code{datasources} argument is not specified
the default set of connections will be used: see \code{\link{datashield.connections_default}}.}
}
\value{
a dataframe with the regression results. These include the estimated coefficients (beta),
their exponents, their standard errors, the upper and lower limits of the 95% confidence intervals,
and the z and p values.
}
\description{
This function runs a distributed Cox Proportional Hazards (PH) Model 
based on the WebDISCO algorithm
}
\details{
For additional details of the WebDISCO algorithm see the paper: 
(\url{https://pmc.ncbi.nlm.nih.gov/articles/PMC5009917/}). The R code was adopted 
from vantage6 and converted to a DataSHIELD function.
}
\author{
Demetris Avraam
}
