% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.Predict.R
\name{ds.Predict}
\alias{ds.Predict}
\title{Creates predictions from regression models fitted with rms on the server side}
\usage{
ds.Predict(
  fit = NULL,
  ...,
  fun = NULL,
  type = c("predictions", "model.frame", "x"),
  conf.int = 0.95,
  conf.type = c("mean", "individual", "simultaneous"),
  ref.zero = FALSE,
  objectname = NULL,
  datasources = NULL
)
}
\arguments{
\item{fit}{character string specifying the name of the fitted model object on the server}

\item{...}{character strings specifying names of predictors to vary}

\item{fun}{character string specifying a transformation for the predictions}

\item{type}{character string specifying type of predictions}

\item{conf.int}{numeric confidence level (0-1)}

\item{conf.type}{character string specifying type of confidence interval}

\item{ref.zero}{logical; if TRUE, predictions are centered at a reference value}

\item{objectname}{character string specifying name for the resulting prediction object}

\item{datasources}{a list of \code{\link{DSConnection-class}} objects obtained after login}
}
\value{
assigns the prediction object to the server side with the specified objectname
}
\description{
This is a wrapper function for rms::Predict that works in the DataSHIELD infrastructure
}
\details{
This function allows users to generate predictions from regression models fitted with rms
on the server side. It supports various types of predictions and confidence intervals.

Server function called: \code{predictDS}
}
\examples{
\dontrun{
  # connecting to the Opal servers
  builder <- DSI::newDSLoginBuilder()
  builder$append(server = "study1",
                 url = "http://192.168.56.100:8080/",
                 user = "administrator",
                 password = "datashield_test&",
                 table = "SURVIVAL.EXPAND_WITH_MISSING1",
                 driver = "OpalDriver")
  logindata <- builder$build()
  connections <- DSI::datashield.login(logins = logindata, assign = TRUE, symbol = "D")
  ds.mice(data = 'D', m = 5, method = 'rf', newobj_df = 'D2', seed = 'fixed', newobj_mids = "imputed_mids")
  ds.asNumeric(x.name = "D2.1$cens",
            newobj = "EVENT",
            datasources = connections)
  ds.asNumeric(x.name = "D2.1$survtime",
            newobj = "SURVTIME",
            datasources = connections)
  ds.asNumeric(x.name = "D2.1$starttime",
            newobj = "STARTTIME",
            datasources = connections)
  ds.asNumeric(x.name = "D2.1$endtime",
            newobj = "ENDTIME",
            datasources = connections)

  ds.datadist(data = 'D2.1', adjust_to = list(age.60 = 'min'))

  ds.useDatadist(datadist = "datadist_D2.1")

  dsSurvivalClient::ds.Surv(time='STARTTIME', time2='ENDTIME',
        event = 'EVENT', objectname='surv_object')

  ds.coxphSLMAassign(formula = 'surv_object ~ age.60',
                        dataName = 'D2.1', objectname = 'cph1', use.rms = TRUE)

  ds.Predict(fit = 'cph1', objectname = 'predictions', fun = "exp", ref.zero = TRUE)

  # When finished, clear the session and logout
  datashield.logout(connections)
}
}
\author{
Xavier Escribà Montagut, 2025
}
