% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_life.table.R
\name{plot_life.table}
\alias{plot_life.table}
\title{Plot Survival Curves}
\usage{
plot_life.table(surv_df, group_col = NULL)
}
\arguments{
\item{surv_df}{A dataframe from survival analysis, must contain 'time', 'surv', 'upper', 'lower' columns.}

\item{group_col}{(Optional) A character string specifying the column name in surv_df for stratified plots.}
}
\value{
A ggplot2 object representing the survival curves.
}
\description{
This function plots the survival curves using ggplot2 and geom_step.
It takes a dataframe resulting from a survival analysis and an optional grouping variable for stratified plots.
}
\examples{
df <- data.frame(time = c(1,2,3), surv = c(0.8, 0.6, 0.4), upper = c(0.9, 0.7, 0.5), lower = c(0.7, 0.5, 0.3), group = c('A', 'B', 'A'))
plot_survival_curves(df, 'group')
}
