% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.datadist.R
\name{ds.datadist}
\alias{ds.datadist}
\title{Creates a server-side datadist object for use with rms package functions}
\usage{
ds.datadist(
  data = NULL,
  adjust_to = NULL,
  objectname = NULL,
  datasources = NULL
)
}
\arguments{
\item{data}{character string specifying the name of a data frame on the server}

\item{adjust_to}{named list where names are variable names and values are either "min", "max", or "mean"
to set the "Adjust to" values for specific variables. For example: 
list(age="min", weight="max", height="mean") will set age's "Adjust to" to its minimum,
weight's to its maximum, and height's to its mean.}

\item{objectname}{character string specifying name of the output object to create on the server}

\item{datasources}{a list of \code{\link{DSConnection-class}} objects obtained after login.
If the \code{datasources} argument is not specified the default set of connections will be used:
see \code{\link{datashield.connections_default}}.}
}
\value{
Assigns a datadist object on the server side containing distribution summaries
}
\description{
Creates a server-side datadist object that contains statistical summaries
of variables for use in plotting regression model fits.
}
\details{
This function creates a server-side datadist object which contains statistical
summaries of variables that are needed for plotting effects in regression models.
The summaries include quantiles, ranges, and other distribution statistics.

Server function called: \code{datadistDS}
}
\examples{
\dontrun{
  ## Version 1.0.0
  
  # connecting to the Opal servers
  builder <- DSI::newDSLoginBuilder()
  builder$append(server = "study1",
                 url = "http://192.168.56.100:8080/",
                 user = "administrator",
                 password = "datashield_test&",
                 table = "SURVIVAL.EXPAND_WITH_MISSING1",
                 driver = "OpalDriver")
  logindata <- builder$build()
  connections <- DSI::datashield.login(logins = logindata, assign = TRUE, symbol = "D")
  ds.mice(data = 'D', m = 5, method = 'rf', newobj_df = 'D2', seed = 'fixed', newobj_mids = "imputed_mids")

  ds.datadist(data = 'D2.1', adjust_to = list(age.60 = 'min'))

  # When finished, clear the session and logout
  datashield.logout(connections)
}

}
\author{
Xavier Escribà Montagut, 2025
}
