% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.plotsurvfit.R
\name{ds.plotsurvfit}
\alias{ds.plotsurvfit}
\title{Performs plotting of privacy preserving survival curves.}
\usage{
ds.plotsurvfit(formula = NULL, dataName = NULL, datasources = NULL)
}
\arguments{
\item{formula}{character string
    specifying the name of survfit object on the server-side created using ds.survfit().
For more information see \strong{Details}.}

\item{dataName}{character string of name of data frame}

\item{datasources}{a list of \code{\link{DSConnection-class}} objects obtained after login. 
If the \code{datasources} argument is not specified
the default set of connections will be used: see \code{\link{datashield.connections_default}}.}
}
\value{
privacy preserving survival curve from the server side environment.
}
\description{
Performs plotting of privacy preserving survival curves.
}
\details{
This is a function that performs plotting of privacy preserving survival curves.

Server function called: \code{plotsurvfitDS}.
}
\examples{
\dontrun{

  ## Version 6
  
  # connecting to the Opal servers

  require('DSI')
  require('DSOpal')
  require('dsBaseClient')

  builder <- DSI::newDSLoginBuilder()
  builder$append(server = "study1", 
                 url = "http://192.168.56.100:8080/", 
                 user = "administrator", password = "datashield_test&", 
                 table = "SURVIVAL.EXPAND_NO_MISSING1", driver = "OpalDriver")
  builder$append(server = "study2", 
                 url = "http://192.168.56.100:8080/", 
                 user = "administrator", password = "datashield_test&", 
                 table = "SURVIVAL.EXPAND_NO_MISSING2", driver = "OpalDriver")
  builder$append(server = "study3",
                 url = "http://192.168.56.100:8080/", 
                 user = "administrator", password = "datashield_test&", 
                 table = "SURVIVAL.EXPAND_NO_MISSING3", driver = "OpalDriver")
  logindata <- builder$build()
  
  connections <- DSI::datashield.login(logins = logindata, assign = TRUE, symbol = "D") 
  
  # make sure that the outcome is numeric 
  ds.asNumeric(x.name = "D$cens",
            newobj = "EVENT",
            datasources = connections)

  ds.asNumeric(x.name = "D$survtime",
            newobj = "SURVTIME",
            datasources = connections)

  dsBaseClient::ds.Surv(time='SURVTIME', event='EVENT', objectname='surv_object')

  dsBaseClient::ds.coxph.SLMA(formula = 'surv_object ~  D$female', 
            dataName = 'D', datasources = connections)
  
  # clear the Datashield R sessions and logout
  datashield.logout(connections)
}

}
\author{
Soumya Banerjee, Tom Bishop, Demetris Avraam, Paul Burton and DataSHIELD technical team (2021).
}
