% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.plotsurvfit.R
\name{ds.plotsurvfit}
\alias{ds.plotsurvfit}
\title{Performs plotting of privacy preserving survival curves in DataSHIELD.}
\usage{
ds.plotsurvfit(
  formula = NULL,
  dataName = NULL,
  fun = NULL,
  ggplot = FALSE,
  datasources = NULL,
  xlab = "",
  ylab = ""
)
}
\arguments{
\item{formula}{character string
    specifying the name of survfit object on the server-side created using ds.survfit().
For more information see \strong{Details}.}

\item{dataName}{character string of name of data frame}

\item{ggplot}{If `TRUE`, use survminer::ggsurvplot to plot the survival curves}
\item{fun}{optional parameter to have an argument. For example, you can pass 'cloglog' for a log-log survival plot.}

\item{datasources}{a list of \code{\link{DSConnection-class}} objects obtained after login. 
If the \code{datasources} argument is not specified
the default set of connections will be used: see \code{\link{datashield.connections_default}}.}

\item{xlab}{X-axis label, a character string. Default value is ''.}

\item{ylab}{Y-axis label, a character string. Default value is ''.}
}
\value{
privacy preserving survival curve from the server side environment.
}
\description{
Performs plotting of privacy preserving survival curves.
}
\details{
This is a server-side function that plots privacy preserving survival curves in DataSHIELD.

Server function called: \code{plotsurvfitDS}.
}
\examples{
\dontrun{

  ## Version 2.0
  
  # connecting to the Opal servers

  require('DSI')
  require('DSOpal')
  require('dsBaseClient')
  library(dsSurvivalClient)

  builder <- DSI::newDSLoginBuilder()
  builder$append(server = "study1", 
                 url = "http://192.168.56.100:8080/", 
                 user = "administrator", password = "datashield_test&", 
                 table = "SURVIVAL.EXPAND_NO_MISSING1", driver = "OpalDriver")
  builder$append(server = "study2", 
                 url = "http://192.168.56.100:8080/", 
                 user = "administrator", password = "datashield_test&", 
                 table = "SURVIVAL.EXPAND_NO_MISSING2", driver = "OpalDriver")
  builder$append(server = "study3",
                 url = "http://192.168.56.100:8080/", 
                 user = "administrator", password = "datashield_test&", 
                 table = "SURVIVAL.EXPAND_NO_MISSING3", driver = "OpalDriver")
  logindata <- builder$build()
  
  connections <- DSI::datashield.login(logins = logindata, assign = TRUE, symbol = "D") 
  
  # make sure that the outcome is numeric 
  ds.asNumeric(x.name = "D$cens",
            newobj = "EVENT",
            datasources = connections)

  ds.asNumeric(x.name = "D$survtime",
            newobj = "SURVTIME",
            datasources = connections)

  dsSurvivalClient::ds.Surv(time='SURVTIME', event='EVENT', objectname='surv_object')

  dsSurvivalClient::ds.coxph.SLMA(formula = 'surv_object ~  D$female', 
            dataName = 'D', datasources = connections)

  dsSurvivalClient::ds.survfit(formula='surv_object~1', objectname='survfit_object')

  dsSurvivalClient::ds.plotsurvfit(formula = 'survfit_object')

  dsSurvivalClient::ds.plotsurvfit(formula = 'survfit_object', fun = 'cloglog')
  
  # clear the Datashield R sessions and logout
  datashield.logout(connections)
}

}
\author{
Soumya Banerjee, Demetris Avraam, Paul Burton, Xavier Escriba-Montagut, Juan Gonzalez and Tom RP Bishop (2022).
}
