% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.Predict.R
\name{ds.Predict}
\alias{ds.Predict}
\title{Creates predictions from regression models fitted with rms on the server side}
\usage{
ds.Predict(
  fit = NULL,
  ...,
  fun = NULL,
  type = c("predictions", "model.frame", "x"),
  conf.int = 0.95,
  conf.type = c("mean", "individual", "simultaneous"),
  ref.zero = FALSE,
  objectname = NULL,
  datasources = NULL
)
}
\arguments{
\item{fit}{character string specifying the name of the fitted model object on the server}

\item{...}{character strings specifying names of predictors to vary}

\item{fun}{character string specifying a transformation for the predictions}

\item{type}{character string specifying type of predictions}

\item{conf.int}{numeric confidence level (0-1)}

\item{conf.type}{character string specifying type of confidence interval}

\item{ref.zero}{logical; if TRUE, predictions are centered at a reference value}

\item{objectname}{character string specifying name for the resulting prediction object}

\item{datasources}{a list of \code{\link{DSConnection-class}} objects obtained after login}
}
\value{
assigns the prediction object to the server side with the specified objectname
}
\description{
This is a wrapper function for rms::Predict that works in the DataSHIELD infrastructure
}
\details{
This function allows users to generate predictions from regression models fitted with rms
on the server side. It supports various types of predictions and confidence intervals.

Server function called: \code{predictDS}
}
\examples{
\dontrun{
  # Assuming you have a fitted model 'fit' on the server
  ds.Predict(fit="fit", age=30:70, sex="both", 
            conf.int=0.95, ref.zero=TRUE, objectname="predictions")
}
}
\author{
Xavier Escribà Montagut, 2025
}
