% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.rcs.R
\name{ds.rcs}
\alias{ds.rcs}
\title{Creates restricted cubic splines using rms package}
\usage{
ds.rcs(x = NULL, knots = 5, objectname = NULL, datasources = NULL)
}
\arguments{
\item{x}{character string specifying the name of the server-side variable to be transformed}

\item{knots}{integer specifying the number of knots for restricted cubic splines. Default is 5}

\item{objectname}{character string specifying the name for the created server-side object}

\item{datasources}{a list of \code{\link{DSConnection-class}} objects obtained after login.
If the \code{datasources} argument is not specified
the default set of connections will be used: see \code{\link{datashield.connections_default}}}
}
\value{
Server-side object containing the restricted cubic spline transformation
}
\description{
Creates server-side restricted cubic splines using rms::rcs.
}
\details{
This function creates restricted cubic spline transformations on the server side
    using the rms package. It is specifically designed for creating non-linear terms
    in survival analysis models.

Server function called: \code{rcsDS}
}
\examples{
\dontrun{
  ## Version 1.0.0
  
  # connecting to the Opal servers
  require('DSI')
  require('DSOpal')
  require('dsBaseClient')
  
  builder <- DSI::newDSLoginBuilder()
  builder$append(server = "study1", 
                url = "http://192.168.56.100:8080/", 
                user = "administrator", password = "datashield_test&", 
                table = "SURVIVAL.EXPAND_NO_MISSING1", driver = "OpalDriver")
  logindata <- builder$build()
  
  connections <- DSI::datashield.login(logins = logindata, assign = TRUE, symbol = "D")
  
  # Create restricted cubic splines for age variable
  dsSurvivalClient::ds.rcs(x = "D$age",
                          knots = 5,
                          objectname = "age_rcs")
  
  # Use the transformed variable in a Cox model
  dsSurvivalClient::ds.coxph.SLMA(formula = "surv_object ~ age_rcs")
  
  # Clear the Datashield R sessions and logout
  datashield.logout(connections)
}

}
\author{
Xavier Escribà Montagut, 2025
}
