% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.rmsTrans.R
\name{ds.rmsTrans}
\alias{ds.rmsTrans}
\title{Creates restricted cubic splines and other transformations using rms package}
\usage{
ds.rmsTrans(
  x = NULL,
  transformation = "rcs",
  parms = NULL,
  objectname = NULL,
  datasources = NULL
)
}
\arguments{
\item{x}{character string specifying the name of the server-side variable to be transformed}

\item{transformation}{character string specifying the type of transformation. Default is "rcs".
Possible values are:
\itemize{
  \item "rcs" - restricted cubic splines
  \item "asis" - linear (untransformed)
  \item "pol" - polynomial
  \item "lsp" - linear spline
  \item "catg" - categorical
  \item "scored" - ordinal scores
  \item "strat" - stratification
  \item "gTrans" - custom transformation function
}}

\item{parms}{parameters specific to the transformation type:
\itemize{
  \item "rcs" - numeric vector specifying knot locations
  \item "asis" - NULL (no parameters needed)
  \item "pol" - integer specifying polynomial degree
  \item "lsp" - numeric vector specifying knot locations for linear spline
  \item "catg" - numeric vector specifying cut points for categories
  \item "scored" - numeric vector specifying scores for ordinal levels
  \item "strat" - numeric vector specifying cut points for stratification
  \item "gTrans" - function to transform variable using custom function
}}

\item{objectname}{character string specifying the name for the created server-side object}

\item{datasources}{a list of \code{\link{DSConnection-class}} objects obtained after login.
If the \code{datasources} argument is not specified
the default set of connections will be used: see \code{\link{datashield.connections_default}}}
}
\value{
\code{rmsTransDS} returns to the client-side a transformed variable for use in
        survival analysis models
}
\description{
Creates server-side transformed variables using rms::rms.trans, primarily for restricted cubic splines.
}
\details{
This function creates transformed variables on the server side using the rms package,
    with a primary focus on restricted cubic splines (rcs). It supports various transformations
    available in the rms package and is particularly useful for creating non-linear terms
    in survival analysis models.

Server function called: \code{rmsTransDS}
}
\examples{
\dontrun{
  ## Version 1.0.0

  # connecting to the Opal servers
  require('DSI')
  require('DSOpal')
  require('dsBaseClient')

  builder <- DSI::newDSLoginBuilder()
  builder$append(server = "study1",
                url = "http://192.168.56.100:8080/",
                user = "administrator", password = "datashield_test&",
                table = "SURVIVAL.EXPAND_WITH_MISSING1", driver = "OpalDriver")
  logindata <- builder$build()

  connections <- DSI::datashield.login(logins = logindata, assign = TRUE, symbol = "D")

  # Examples of different transformations:

  ds.mice(data = 'D', m = 5, method = 'rf', newobj_df = 'D2', seed = 'fixed', newobj_mids = "imputed_mids")
  ds.rmsTrans(x = "D2.1$age.60",
              transformation = "rcs",
              parms = c(20, 40, 60, 80),
              objectname = "age_rcs")
  ds.rmsTrans(x = "D2.1$bmi",
                 transformation = "pol",
                 parms = 2,
              objectname = "bmi_poly")
  # Clear the Datashield R sessions and logout
  datashield.logout(connections)
}

}
\author{
Xavier Escribà Montagut, 2025
}
