% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.useDatadist.R
\name{ds.useDatadist}
\alias{ds.useDatadist}
\title{Sets a server-side datadist object as the global option}
\usage{
ds.useDatadist(datadist = NULL, datasources = NULL)
}
\arguments{
\item{datadist}{character string specifying the name of a datadist object on the server}

\item{datasources}{a list of \code{\link{DSConnection-class}} objects obtained after login.
If the \code{datasources} argument is not specified the default set of connections will be used:
see \code{\link{datashield.connections_default}}.}
}
\value{
A message indicating the success or failure of the operation
}
\description{
Sets a previously created datadist object as the global option for rms plotting functions
}
\details{
This function sets a datadist object previously created with ds.datadist
as the global option for use with rms plotting functions.

Server function called: \code{useDatadistDS}
}
\examples{
\dontrun{
  ## Version 1.0.0
  
  # connecting to the Opal servers
  builder <- DSI::newDSLoginBuilder()
  builder$append(server = "study1", 
                url = "http://192.168.56.100:8080/", 
                user = "administrator", password = "datashield_test&", 
                table = "SURVIVAL.EXPAND_NO_MISSING1", driver = "OpalDriver")
  logindata <- builder$build()
  
  connections <- DSI::datashield.login(logins = logindata, assign = TRUE, symbol = "D")
  
  # Create datadist object
  ds.datadist(data = "D", objectname = "dd")
  
  # Set the datadist object as the global option
  ds.useDatadist(datadist = "dd")
  
  # Now rms plotting functions will use this datadist object
  
  # Clear the Datashield R sessions and logout
  datashield.logout(connections)
}

}
\author{
Xavier Escribà Montagut, 2025
}
